requirements.txt
appdirs==1.4.4
attrs==20.3.0
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.2.1
defusedxml==0.6.0
Flask==1.1.2
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
PyJWT==1.4.2
lxml==4.6.2
MarkupSafe==1.1.1
pycparser==2.20
pytz==2020.5
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1
zeep==4.0.0
